feat: pool api crate
