+++ crates/core/src/service/mod.rs
@@ -5,8 +5,8 @@ use std::task::{Context, Poll};
 use block_producer::BlockProductionError;
 use futures::stream::StreamExt;
 use katana_executor::ExecutorFactory;
-use katana_pool::ordering::PoolOrd;
-use katana_pool::pending::PendingTransactions;
+use katana_pool::PendingTransactions;
+use katana_pool::PoolOrd;
 use katana_pool::{TransactionPool, TxPool};
 use katana_primitives::transaction::ExecutableTxWithHash;
 use tracing::{error, info};
