+++ crates/pool/pool-api/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+edition.workspace = true
+license.workspace = true
+name = "katana-pool-api"
+version.workspace = true
+
+[dependencies]
+katana-primitives.workspace = true
+
+futures.workspace = true
+parking_lot.workspace = true
+thiserror.workspace = true
+tokio = { workspace = true, features = [ "sync" ] }
+
+[dev-dependencies]
+futures-util.workspace = true
