+++ crates/pool/pool-api/src/validation.rs
@@ -1,7 +1,10 @@
 use katana_primitives::class::ClassHash;
 use katana_primitives::contract::{ContractAddress, Nonce};
 use katana_primitives::execution::Resource;
+use katana_primitives::transaction::TxHash;
 use katana_primitives::Felt;
+
+use crate::PoolTransaction;
 
 // TODO: figure out how to combine this with ExecutionError
 #[derive(Debug, thiserror::Error)]
@@ -119,4 +122,63 @@ pub enum InsufficientFundsError {
     /// data gas bound. But on 0.14.0, it is required to set all bounds.
     #[error("{error}")]
     ResourceBoundsExceedFunds { error: String },
+}
+
+// outcome of the validation phase. the variant of this enum determines on which pool
+// the tx should be inserted into.
+#[derive(Debug)]
+pub enum ValidationOutcome<T> {
+    /// tx that is or may eventually be valid after some nonce changes.
+    Valid(T),
+
+    /// tx that will never be valid, eg. due to invalid signature, nonce lower than current, etc.
+    Invalid { tx: T, error: InvalidTransactionError },
+
+    /// tx that is dependent on another tx ie. when the tx nonce is higher than the current account
+    /// nonce.
+    Dependent {
+        tx: T,
+        /// The nonce that the tx is using.
+        tx_nonce: Nonce,
+        /// The current nonce of the sender's account.
+        current_nonce: Nonce,
+    },
+}
+
+#[derive(Debug, thiserror::Error)]
+#[error("{error}")]
+pub struct Error {
+    /// The hash of the transaction that failed validation.
+    pub hash: TxHash,
+    /// The actual error object.
+    pub error: Box<dyn std::error::Error>,
+}
+
+impl Error {
+    pub fn new(hash: TxHash, error: Box<dyn std::error::Error>) -> Self {
+        Self { hash, error }
+    }
+}
+
+pub type ValidationResult<T> = Result<ValidationOutcome<T>, Error>;
+
+/// A trait for validating transactions before they are added to the transaction pool.
+pub trait Validator {
+    type Transaction: PoolTransaction;
+
+    /// Validate a transaction.
+    ///
+    /// The `Err` variant of the returned `Result` should only be used to represent unexpected
+    /// errors that occurred during the validation process ie, provider
+    /// [error](katana_provider::error::ProviderError), and not for indicating that the
+    /// transaction is invalid. For that purpose, use the [`ValidationOutcome::Invalid`] enum.
+    fn validate(&self, tx: Self::Transaction) -> ValidationResult<Self::Transaction>;
+
+    /// Validate a batch of transactions.
+    fn validate_all(
+        &self,
+        txs: Vec<Self::Transaction>,
+    ) -> Vec<ValidationResult<Self::Transaction>> {
+        txs.into_iter().map(|tx| self.validate(tx)).collect()
+    }
 }
